[DDW-330] Fix withdrawal param for createTransaction and getTransactionFee
